Move LoadingSpinner to admin's components
