+++ src/pages/Admin/Home/Dashboard/TaskDetail/Company/CompanyDetailContent/container.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent } from "react";
-import { LoadingSpinner } from "../../LoadingSpinner";
+import { LoadingSpinner } from "../../../../../components/LoadingSpinner";
 import { CompanyDetailContent } from "./component";
 import { useCompanyByUuid } from "$hooks/queries";
