+++ src/pages/Admin/Home/Dashboard/TaskDetail/JobApplication/DetailContent/container.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from "react";
 import { JobApplicationDetailContent } from "./component";
 import { useApplicantByUuid, useCompanyOfferByUuid } from "$hooks/queries";
 import { IContainerProps } from "./interfaces";
-import { LoadingSpinner } from "../../LoadingSpinner";
+import { LoadingSpinner } from "../../../../../components/LoadingSpinner";
 
 export const JobApplicationDetailContentContainer: FunctionComponent<IContainerProps> = ({
   applicantUuid,
