+++ src/pages/Admin/Home/Dashboard/TaskDetail/Offer/OfferDetailContent/component.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from "react";
 import { OfferDetail } from "$components/OfferDetail";
 import { useOfferByUuid } from "$hooks/queries";
-import { LoadingSpinner } from "../../LoadingSpinner";
+import { LoadingSpinner } from "../../../../../components/LoadingSpinner";
 
 export const OfferDetailContent: FunctionComponent<IOfferDetailContentProps> = ({
   offerUuid,
